requirements.txt
behave==1.2.6
selenium==4.1.0
pyyaml==6.0
webdriver-manager
